requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.0
streamlit>=1.28.0
anthropic>=0.18.0
langchain-anthropic>=0.1.0
openai>=1.12.0
python-dotenv>=1.0.0
langchain>=0.1.0
pyyaml>=6.0.0

